fix: light, lightUser 분리
